totally reworked the prompts. Woo!
