+++ generators/module/templates/module.hbs
@@ -4,6 +4,6 @@ layout: module.hbs
 module: true
 ---
 
-<div class="<%= moduleName %>" data-module="<%= moduleName %>">
+<div class="<%= moduleName %>">
 	<%= moduleName %>
 </div>
