+++ generators/new/templates/_bower.json
