+++ generators/new/templates/_package.json
