+++ generators/new/templates/bowerrc
