+++ generators/new/templates/gitignore
