+++ generators/new/templates/jshintrc
