+++ generators/new/templates/rocket.png
