+++ generators/partial/templates/partial.hbs
@@ -1,3 +1,3 @@
-<div class="<%= partialName %>" data-partial="<%= partialName %>">
+<div class="<%= partialName %>">
 	<%= partialName %>
 </div>
