+++ generators/template/templates/template.hbs
@@ -2,6 +2,6 @@
 title: <%= templateName %>
 template: true
 ---
-<div class="<%= templateName %>" data-template="<%= templateName %>">
+<div class="<%= templateName %>">
 	<%= templateName %>
 </div>
